[r] Generate DRS path for DCP/1 matrices (#2553)
